Extract chat send module
