+++ service-worker.js
@@ -77,6 +77,7 @@ const APP_SHELL = [
   '/js/chat-discussion.js',
   '/js/chat-onboarding.js',
   '/js/chat-render.js',
+  '/js/chat-send.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
